[fix] 배틀에 사용할 dto 수정
- MonsterDTO.MonsterBattleInfo의 skill은 id가 아닌 객체를 저장하도록 함
