+++ BE/src/main/java/cuketmon/battle/dto/BattleDTO.java
@@ -1,6 +1,6 @@
 package cuketmon.battle.dto;
 
-import cuketmon.type.Type;
+import cuketmon.monster.dto.MonsterDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -11,44 +11,9 @@ public class BattleDTO {
     @Getter
     @AllArgsConstructor
     public static class Team {
-        private Integer battleId;
         private String trainerName;
-        private Monster monster;
+        private MonsterDTO.MonsterBattleInfo monster;
         private Boolean turn;
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class Monster {
-        private String name;
-        private String image;
-        private Integer affinity;
-
-        private Integer hp;
-        private Integer speed;
-        private Integer attack;
-        private Integer defence;
-        private Integer specialAttack;
-        private Integer specialDefence;
-
-        private Type type1;
-        private Type type2;
-
-        private Skill skill1;
-        private Skill skill2;
-        private Skill skill3;
-        private Skill skill4;
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class Skill {
-        private Type type;
-        private String damageClass;
-        private Integer accuracy;
-        private String name;
-        private Integer power;
-        private Integer pp;
     }
 
 }
